lb_marketing_api/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
debugpy==1.8.11
alembic==1.13.2
azure-functions==1.18.0
